// src/api/generated/model/mgmtRolloutGroupResponseBody.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtRolloutCondition } from './mgmtRolloutCondition';
import type { MgmtRolloutSuccessAction } from './mgmtRolloutSuccessAction';
import type { MgmtRolloutErrorAction } from './mgmtRolloutErrorAction';
import type { MgmtRolloutGroupResponseBodyTotalTargetsPerStatus } from './mgmtRolloutGroupResponseBodyTotalTargetsPerStatus';
import type { Links } from './links';

export interface MgmtRolloutGroupResponseBody {
  /** Entity was originally created by (User, AMQP-Controller, anonymous etc.) */
  readonly createdBy?: string;
  /** Entity was originally created at (timestamp UTC in milliseconds) */
  readonly createdAt?: number;
  /** Entity was last modified by (User, AMQP-Controller, anonymous etc.) */
  readonly lastModifiedBy?: string;
  /** Entity was last modified at (timestamp UTC in milliseconds) */
  readonly lastModifiedAt?: number;
  name: string;
  description?: string;
  successCondition?: MgmtRolloutCondition;
  successAction?: MgmtRolloutSuccessAction;
  errorCondition?: MgmtRolloutCondition;
  errorAction?: MgmtRolloutErrorAction;
  /** The name of the entity */
  targetFilterQuery?: string;
  /** Percentage of remaining and matching targets that should be added to this group */
  targetPercentage?: number;
  /** (Available with user consent flow active) If the confirmation is required for this rollout group. Confirmation is required per default */
  confirmationRequired?: boolean;
  /** Rollouts id */
  id: number;
  /** If the rollout group is dynamic */
  dynamic?: boolean;
  /** The status of this rollout */
  status?: string;
  /** The total targets of a rollout */
  totalTargets?: number;
  /** The total targets per status */
  totalTargetsPerStatus?: MgmtRolloutGroupResponseBodyTotalTargetsPerStatus;
  _links?: Links;
}